Don't log UI style if the app is always in light mode or always in dark mode
It is possible to put an app permanently into light mode (or dark mode) using the Info.plist key `UIUserInterfaceStyle`.
If this is set to either of these values, the user's dark mode preference has no effect and the app will report the override interface style. In this case, don't include it in the statistics.